# Fix and optimize rmsnorm OpenCL implementation, inspired by GGML opecl kernel.