--- a/nntrainer/tensor/cl_operations/blas_kernels_templates.h
+++ b/nntrainer/tensor/cl_operations/blas_kernels_templates.h
@@ -334,11 +334,6 @@ inline static void rmsnorm_cl_internal(ClContext::SharedPtrClKernel kernel,
     static_cast<ClContext *>(Engine::Global().getRegisteredContext("gpu"));
 
   if (use_svm) {
-    blas_cc->command_queue_inst_.enqueueSVMMap(const_cast<T *>(input), size_in,
-                                               true);
-    blas_cc->command_queue_inst_.enqueueSVMMap(const_cast<T *>(gamma),
-                                               size_gamma, true);
-
     if (!kernel->SetKernelSVMArguments(0, input)) {
       return;
     }
@@ -382,19 +377,21 @@ inline static void rmsnorm_cl_internal(ClContext::SharedPtrClKernel kernel,
   if (!kernel->SetKernelArguments(5, &width, sizeof(int))) {
     return;
   }
-  const int work_groups_count[3] = {static_cast<int>(height) * 32, 1, 1};
+#ifdef __ANDROID__
+  constexpr int SUBGROUP_SIZE = 64;
+#else
+  constexpr int SUBGROUP_SIZE = 32;
+#endif
+  const int work_groups_count[3] = {static_cast<int>(height) * SUBGROUP_SIZE, 1,
+                                    1};
 
-  const int work_group_size[3] = {32, 1, 1}; // test-value
+  const int work_group_size[3] = {SUBGROUP_SIZE, 1, 1};
   if (!blas_cc->command_queue_inst_.DispatchCommand(kernel, work_groups_count,
                                                     work_group_size)) {
     return;
   }
 
-  if (use_svm) {
-    if (!blas_cc->command_queue_inst_.enqueueSVMUnmap(result)) {
-      return;
-    }
-  } else {
+  if (!use_svm) {
     auto &clbuffInstance = ClBufferManager::Global();
     if (!clbuffInstance.getOutBufferA()->ReadDataRegion(
           blas_cc->command_queue_inst_, size_in, result)) {