This commit fixes and optimizes RMSNorm implementation of OpenCL kernel. Details:

1. `rmsnorm_cl` function is added to `blas_kernels.h`
2. RMSNormLayerCL now only calls `rmsnorm_cl` instead of setting up the kernel
3. `rmsnorm_cl` optionally uses SVM
4. The implementation is change inspired by GGML implementation - there's one subgroup per workgroup and each workgroup cooperatively calculates sum of elements in each row and then normalizes the row. Performance was tested for `3072x3072` tensor.
5. Currently there's an additional, commented implementation of kernel which uses 64 size workgroup (2 subgroups). On my machine (Intel UHD Graphics) there's no difference in performance between them, but on something other there can be.

| CPU      | GPU Before | GPU After         |
|----------|------------|-------------------|
| 25-40 ms | ~200 ms    | 3-5 ms (with SVM) | 

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
6. Run test:   [X]Passed [ ]Failed [ ]Skipped